Move `QualifiersActionTest` tests to other existing test files
